front-end/src/app/modules/modal/services/person-register/person-register.service.ts
import { Injectable } from '@angular/core';
import { HttpClient, HttpHeaders } from '@angular/common/http';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root'
})
export class PersonRegisterService {

  private url = 'http://localhost:3000/create';

  constructor(private http: HttpClient) { }

  createPerson(data: any): Observable<any> {
    const token = localStorage.getItem('token');
    const headers = new HttpHeaders({
      'Content-Type': 'application/json',
      Authorization: `Bearer ${token}`,
    });

    return this.http.post(this.url, data, { headers, observe: 'response' });
  }
}
